requirements.txt
streamlit
chromadb
sentence-transformers
python-dotenv
groq
numpy
pandas
plotly